requirements.txt
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
asynctest==0.13.0
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
Django==3.2.20
django-admin-interface==0.26.0
django-colorfield==0.9.0
django-multiselectfield==0.1.12
django-phonenumber-field==7.1.0
django-twilio==0.13.2
djangorestframework==3.14.0
EasyProcess==1.1
entrypoint2==1.1
Flask==2.2.5
frozenlist==1.3.3
idna==3.4
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
MouseInfo==0.1.3
mss==7.0.1
multidict==6.0.4
phonenumbers==8.13.19
Pillow==9.5.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyJWT==1.7.1
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
python-slugify==8.0.1
pytweening==1.0.7
pytz==2023.3
pywhatkit==5.4
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
text-unidecode==1.3
twilio==6.63.2
typing_extensions==4.7.1
urllib3==1.26.6
uuid==1.30
Werkzeug==2.2.3
wikipedia==1.4.0
yarl==1.9.2
zipp==3.15.0
